refactor(serde_yml): :art: refactor: remove dependencies and unecessay functions
Changes based on @QuadnucYard PR

- Changes based on https://github.com/sebastienrousseau/serde_yml/compare/master...QuadnucYard:serde_yml:relex-deps-radical
