+++ .editorconfig
@@ -1,5 +1,4 @@
-# Copyright © 2023 serde_yml. All rights reserved.
-# SPDX-License-Identifier: Apache-2.0 OR MIT
+# SPDX-License-Identifier: MIT
 
 # EditorConfig helps developers define and maintain consistent
 # coding styles between different editors and IDEs
