+++ src/lib.rs
@@ -37,7 +37,7 @@
 //!
 //! ## Rust Version Compatibility
 //!
-//! This library is compatible with Rust 1.60 and above.
+//! This library is compatible with Rust 1.56.0 and above.
 //!
 //! ## Installation
 //!
@@ -145,9 +145,6 @@ pub mod libyml;
 /// The `loader` module contains the `Loader` type for YAML loading.
 pub mod loader;
 
-/// The `macros` module contains functions for generating macros.
-pub mod macros;
-
 /// The `mapping` module contains the `Mapping` type for YAML mappings.
 pub mod mapping;
 
@@ -159,9 +156,6 @@ pub mod number;
 
 /// The `ser` module contains the library's YAML serializer.
 pub mod ser;
-
-/// The `utilities` module contains utility functions for the library.
-pub mod utilities;
 
 /// The `value` module contains the `Value` type for YAML values.
 pub mod value;
