--- src/macros/macro_from_number.rs
@@ -1,13 +1,0 @@
-/// Macro `from_number` to implement conversion from number types to `Value`.
-#[macro_export]
-macro_rules! from_number {
-    ($($ty:ident)*) => {
-        $(
-            impl From<$ty> for Value {
-                fn from(n: $ty) -> Self {
-                    Value::Number(n.into())
-                }
-            }
-        )*
-    };
-}
