--- src/macros/mod.rs
@@ -1,21 +1,0 @@
-/// The `directory_macros` module contains macros related to directory
-/// operations.
-pub mod macro_directory;
-
-/// The `file_macros` module contains macros related to file operations.
-pub mod macro_file;
-
-/// The `from_number_macros` module contains macros related to converting
-pub mod macro_from_number;
-
-/// The `partialeq_numeric_macros` module contains macros related to the `PartialEq` trait for primitive numeric types and `Value`.
-pub mod macro_partialeq_numeric;
-
-/// The `macro_get_field` module contains macros related to retrieving field values from a JSON file.
-pub mod macro_get_field;
-
-/// The `nested_enum_serde` module contains macros related to serializing and deserializing nested enums.
-pub mod macro_nested_enum_serde;
-
-/// The `replace_placeholder_macros` module contains macros related to replacing placeholders in a line with values from parameters.
-pub mod macro_replace_placeholder;
