+++ src/modules/path.rs
@@ -1,4 +1,3 @@
-use serde::Serialize;
 use std::fmt::{self, Display};
 
 /// `Path` represents the path to the current value in the input, like `dependencies.serde.typo1`.
@@ -13,7 +12,7 @@ use std::fmt::{self, Display};
 /// - `Map`: Represents a map (object) path with a reference to the parent path and a key.
 /// - `Alias`: Represents an alias path with a reference to the parent path.
 /// - `Unknown`: Represents an unknown path with a reference to the parent path.
-#[derive(Copy, Clone, Debug, PartialEq, Serialize)]
+#[derive(Copy, Clone, Debug, PartialEq)]
 pub enum Path<'a> {
     /// Represents the root path.
     Root,
