--- src/utilities/mod.rs
@@ -1,2 +1,0 @@
-/// The `directory` module contains functions related to directory operations.
-pub mod directory;
