--- tests/macros/mod.rs
@@ -1,20 +1,0 @@
-/// The `test_macro_directory` module contains tests for the directory module.
-pub mod test_macro_directory;
-
-/// The `test_macro_file` module contains tests for the file module.
-pub mod test_macro_file;
-
-/// The `test_macro_nested_enum_serde` module contains tests for the nested enum serde module.
-pub mod test_macro_nested_enum_serde;
-
-/// The `test_macro_partialeq_numeric` module contains tests for the partialeq_numeric module.
-pub mod test_macro_partialeq_numeric;
-
-/// The `test_macro_replace_placeholder` module contains tests for the utility module.
-pub mod test_macro_replace_placeholder;
-
-/// The `test_macro_get_field` module contains tests for the get_field module.
-pub mod test_macro_get_field;
-
-/// The `test_macro_from_number` module contains tests for the from_number module.
-pub mod test_macro_from_number;
