+++ tests/mod.rs
@@ -1,14 +1,8 @@
 /// This module contains the tests for the `libyml` module.
 pub mod libyml;
 
-/// This module contains the tests for the `macros` module.
-pub mod macros;
-
 /// This module contains the tests for the `modules` module.
 pub mod modules;
-
-/// This module contains the tests for the `utilities` module.
-pub mod utilities;
 
 /// This module contains the tests for the `value` module.
 pub mod value;
