--- tests/utilities/mod.rs
@@ -1,2 +1,0 @@
-/// The `test_directory.rs` module contains tests for the directory module.
-pub mod test_directory;
